RepTim (first version)
This package aims to perform the differential analysis between replication timing profiles (biological conditions) in order to detect modified regions. Then, it localizes genes in regions of interest (e.g. modified regions) and tests if biological pathways are enriched. It does also a Fisher's exact test to find transcription factors for wich more genes (targets) are found in the modified regions in comparison with non-modified regions.